Move globalStyles
